feat: dual-venvs for faster/cheaper env creation
