+++ integration_tests/projects/013_structured_hooks/scripts/environment_type.py
@@ -2,7 +2,7 @@ import sys
 from pathlib import Path
 
 from fal.typing import *
-from fal.packages.environment import _BASE_VENV_DIR
+from fal.packages.environments.virtual_env import _BASE_VENV_DIR
 from _fal_testing import create_dynamic_artifact
 
 # To determine whether this is a fal-created environment or not
