--- src/fal/packages/__init__.py
@@ -1,1 +1,0 @@
-from fal.packages.environment import BaseEnvironment, VirtualPythonEnvironment
